wackozacko isaac.doub@gmail.com 30/04/2011
Created Base64FileEncoder.java. This class takes a
file or a directory and will convert it to base 64
and output it to a directory 'base64'.

Command line parameters are as follows:

-h  insert encodings into html (data/png;base64 et
c). Will hopefully call a future function which wi
ll actually replace the url where the image was fo
und with the base64 encoding.

-s  write all encodings to a single file instead o
f individual files. Titles each chunk with the ori
ginal filename.

-lt <filesize>  this option sets a maximum to the
filesize that will be encoded in KB. If an image i
s larger than this size it is simply skipped.

-c  this only encodes files if the new filesize is
less than the original filesize * 1.35 . Unfortuna
tely because base64 is unoptimized the new filesiz
es are always greater than the original filesize *
1.35 SO DON'T USE THIS. I have left it in as I am
unsure whether GZipping the text files will improv
e this as of yet. If that is the case, I can chang
e the factor from 1.35 to perhaps greater and then
it will be a useful feature.